pyproject.toml
[project]
name = "ffmpeg-binaries"
authors = [{name="Matteo Dell'Acqua", email="dellacqua.matteo99@gmail.com"}]
maintainers = [{name="Matteo Dell'Acqua", email="dellacqua.matteo99@gmail.com"}]
description = "A static build of ffmpeg for Python"
license = {file="LICENSE"}
requires-python = ">=3.9"
keywords = ["ffmpeg", "static"]
dynamic = ["version"]
readme = "README.md"
dependencies = [
    'requests',
    'patool'
]
classifiers = [
    'License :: OSI Approved :: MIT License',
    'Programming Language :: Python :: 3 :: Only',
    'Programming Language :: Python :: 3.9',
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
    'Programming Language :: Python :: 3.12',
    'Programming Language :: Python :: 3.13'
]

[project.scripts]
ffmpeg = "ffmpeg_binaries.executable:_run"

[build-system]
requires = ["setuptools>=43.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.dynamic]
version = {attr = "ffmpeg_binaries.__version__"}

[tool.setuptools.package-data]
"*" = ["*.json"]
